luxury/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.15
django-ckeditor==6.3.2
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
huey==2.4.5